New sensor testing (#25)
Update temperature and infrared sensors based on testing.

+++ pybricks/nxtdevices/__stub/__temperaturesensor.py
@@ -13,11 +13,11 @@ class TemperatureSensor:
         if port == Port.A or port == Port.B or port == Port.C or port == Port.D:
             raise ValueError("Sensors must use Port S1, S2, S3, or S4.")
 
-    def temperature(self) -> int:
+    def temperature(self) -> float:
         """
         Measures the temperature.
 
         Returns:
             Measured temperature in degrees Celsius.
         """
-        return 0
+        return 0.0
