+++ test/nxtdevices/test_temperaturesensor.py
@@ -21,5 +21,5 @@ class TestTemperatureSensor(TestCase):
     def test_temperature(self):
         sensor = TemperatureSensor(Port.S1)
         result = sensor.temperature()
-        self.assertIsInstance(result, int)
-        self.assertEqual(0, result)
+        self.assertIsInstance(result, float)
+        self.assertEqual(0.0, result)
